combine schemas.py and lcs.py into auth.py
