+++ src/flaskapp/api.py
@@ -24,7 +24,7 @@ from src.teams.unify.team_reject import team_reject
 # from src.matching.team_recommendations import get_team_recommendations
 
 from src.flaskapp.util import format_string
-from src.flaskapp.schemas import ensure_user_logged_in
+from src.flaskapp.auth import authenticate
 
 app = Flask(__name__)
 
@@ -36,7 +36,7 @@ def index():
 
 
 @app.route("/users", methods=["GET", "POST"])
-@ensure_user_logged_in
+@authenticate
 def users(email):
     email = None  # unused
 
@@ -85,7 +85,7 @@ def users(email):
 
 
 @app.route("/users/profile", methods=["GET", "PUT"])
-@ensure_user_logged_in
+@authenticate
 def single_user(email):
     if request.method == "GET":
         # Retrieve a single user
@@ -106,7 +106,7 @@ def single_user(email):
 
 
 @app.route("/teams", methods=["GET", "POST"])
-@ensure_user_logged_in
+@authenticate
 def teams(email):
     email = None  # unused
 
@@ -141,7 +141,7 @@ def teams(email):
 
 
 @app.route("/teams/<team_id>", methods=["GET", "PUT"])
-@ensure_user_logged_in
+@authenticate
 def single_team(email, team_id):
     if request.method == "GET":
         return get_team_profile(email, team_id)
@@ -151,13 +151,13 @@ def single_team(email, team_id):
 
 
 @app.route("/teams/<team_id>/complete", methods=["PUT"])
-@ensure_user_logged_in
+@authenticate
 def mark_team_complete(email, team_id):
     return team_complete(email, team_id)
 
 
 @app.route("/teams/<team_id>/leave", methods=["PUT"])
-@ensure_user_logged_in
+@authenticate
 def leave(email, team_id):
     return user_leave(email, team_id)
 
@@ -166,7 +166,7 @@ def leave(email, team_id):
 
 
 @app.route("/teams/<team_id>/invite", methods=["POST"])
-@ensure_user_logged_in
+@authenticate
 def invite(email, team_id):
     # NOTE team1 -inviting-> team2 (invite another team)
     team1_name = team_id
@@ -178,7 +178,7 @@ def invite(email, team_id):
 
 
 @app.route("/teams/<team_id>/confirm", methods=["POST"])
-@ensure_user_logged_in
+@authenticate
 def confirm(email, team_id):
     # NOTE team1 -confirms-> team2 (confirm an invite)
     team1_name = team_id
@@ -190,7 +190,7 @@ def confirm(email, team_id):
 
 
 @app.route("/teams/<team_id>/rescind", methods=["POST"])
-@ensure_user_logged_in
+@authenticate
 def rescind(email, team_id):
     # NOTE team1 -rescind-> team2 (rescind an invite)
     team1_name = team_id
@@ -202,7 +202,7 @@ def rescind(email, team_id):
 
 
 @app.route("/teams/<team_id>/reject", methods=["POST"])
-@ensure_user_logged_in
+@authenticate
 def reject(email, team_id):
     # NOTE team1 -reject-> team2 (rejecting an invite)
     team1_name = team_id
@@ -217,7 +217,7 @@ def reject(email, team_id):
 
 
 @app.route("/matches/<team_id>", methods=["GET"])
-@ensure_user_logged_in
+@authenticate
 def team_recommendations(email, team_id):
     # WIP
     # return get_team_recommendations(email, team_id)
